tests/openfaas/image-recog/requirements.txt
numpy==1.26.4
torch==1.12.1
torchvision==0.13.1
Pillow==9.2.0
